feat: add transition-group dependency and refactor link styles
- Added @solid-primitives/transition-group to dependencies in package.json and bun.lock.
- Removed legacy CSS file theme-link.css and updated link styles to use Tailwind utilities for modern animations.
- Refactored link components to utilize new styles and removed ClientMotion components from several sections for improved performance.
