--- legacy-css/theme-link.css
@@ -1,167 +1,0 @@
-/* ========================================
-   CSS Custom Properties (Variables)
-   ======================================== */
-
-:root {
-  --theme-link-primary: #007bff;
-  --theme-link-secondary: #6c757d;
-  --theme-link-hover: #0056b3;
-  --theme-link-active: #004085;
-  --theme-link-visited: #6f42c1;
-  --theme-link-transition-duration: 0.2s;
-  --theme-link-border-radius: 4px;
-  --theme-link-padding: 8px 16px;
-  --theme-link-font-weight: 500;
-  
-  /* Animation variables */
-  --underline-duration: 280ms;
-  --underline-easing: cubic-bezier(0.23, 1, 0.320, 1);
-  --link-transition-duration: 0.3s;
-  --link-transition-easing: cubic-bezier(0.4, 0, 0.2, 1);
-  --link-exit-duration: 0.25s;
-  --link-exit-easing: cubic-bezier(0.5, 0, 0.75, 0);
-  --link-enter-duration: 0.6s;
-  --link-enter-easing: cubic-bezier(0.16, 1, 0.3, 1);
-}
-
-/* ========================================
-   @keyframes Declarations
-   ======================================== */
-
-@keyframes theme-link-pulse {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.8;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-@keyframes theme-link-fade-in {
-  from {
-    opacity: 0;
-    transform: translateY(-2px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@keyframes theme-link-underline-expand {
-  from {
-    width: 0;
-  }
-  to {
-    width: 100%;
-  }
-}
-
-@keyframes theme-link-enter {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@keyframes theme-link-exit {
-  from {
-    opacity: 1;
-    transform: translateY(0) scale(1);
-    color: hsl(var(--accent));
-  }
-  to {
-    opacity: 0.8;
-    transform: translateY(-2px) scale(0.98);
-    color: hsl(var(--muted-foreground));
-  }
-}
-
-@keyframes underline-grow {
-  from {
-    transform: scaleX(0);
-  }
-  to {
-    transform: scaleX(1);
-  }
-}
-
-@keyframes underline-shrink {
-  from {
-    transform: scaleX(1);
-  }
-  to {
-    transform: scaleX(0);
-  }
-}
-
-/* ========================================
-   Theme-link Class Selectors and Descendant Rules
-   ======================================== */
-
-.theme-link {
-  color: hsl(var(--accent));
-  font-weight: 500;
-  text-decoration: none;
-  transition: all 0.15s;
-  position: relative;
-  display: inline;
-  opacity: 1;
-}
-
-.theme-link:hover {
-  color: hsl(var(--accent));
-  transition: all var(--link-transition-duration) var(--link-transition-easing);
-}
-
-.theme-link:not(:hover) {
-  transition: all var(--link-exit-duration) var(--link-exit-easing);
-}
-
-.theme-link::before {
-  content: '';
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  height: 1px;
-  transform: scaleX(0);
-  transform-origin: left;
-  transition: transform var(--underline-duration) var(--underline-easing);
-}
-
-.theme-link:hover::before {
-  animation: underline-grow var(--underline-duration) var(--underline-easing) forwards;
-}
-
-.theme-link:not(:hover)::before {
-  animation: underline-shrink var(--underline-duration) var(--underline-easing) forwards;
-}
-
-.theme-link.with-arrow::before {
-  right: 1.5rem;
-  background: linear-gradient(to right, hsl(var(--accent)) 70%, transparent 100%);
-}
-
-.theme-link:not(.with-arrow)::before {
-  right: 0;
-  background-color: hsl(var(--accent));
-}
-
-.theme-link.with-arrow::after {
-  position: relative;
-  z-index: 1;
-  content: " ↗";
-  margin-right: 0.125rem;
-}
-
-/* Keep accent color variant */
-.theme-link.keep-accent:hover {
-  color: hsl(var(--accent));
-}
