+++ package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@motionone/solid": "^10.16.4",
+    "@solid-primitives/transition-group": "^1.1.2",
     "@solidjs/router": "^0.15.0",
     "@solidjs/start": "^1.1.0",
     "@tanstack/solid-query": "^5.83.0",
