+++ src/modules/sections/about-section.tsx
@@ -1,14 +1,8 @@
-import { ClientMotion } from "~/components/ui/client-motion";
 import { Text } from "~/components/primitives/text";
 
 export function AboutSection() {
   return (
-    <ClientMotion
-      as="div"
-      initial={{ opacity: 0, y: 20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.6, delay: 0.1, easing: [0.4, 0.0, 0.2, 1] }}
-    >
+    <div>
       <Text as="p" variant="body">
         With extensive experience in{" "}
         <Text as="span" variant="highlight">
@@ -20,6 +14,6 @@ export function AboutSection() {
         </Text>
         .
       </Text>
-    </ClientMotion>
+    </div>
   );
 }
