+++ src/modules/sections/timezone-section.tsx
@@ -1,4 +1,4 @@
-import { ClientMotion } from "~/components/ui/client-motion";
+import { For } from "solid-js";
 import { TIMEZONE_INFO, useCurrentTime } from "~/modules/time";
 import { getParagraphClass } from "~/cms";
 
@@ -11,44 +11,29 @@ export function TimezoneSection(props: TProps) {
   const currentTime = useCurrentTime();
 
   return (
-    <Motion.p 
-      class={`text-foreground ${getParagraphClass('body')}`}
-      initial={{ opacity: 0, y: 20 }}
-      animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.6, delay: 0.4, easing: [0.4, 0.0, 0.2, 1] }}
-    >
+    <p class={`text-foreground ${getParagraphClass('body')}`}>
       My current timezone is{" "}
       <span class="dashed-highlight">
         {TIMEZONE_INFO.offset}
       </span>{" "}
       which includes countries like{" "}
-      {TIMEZONE_INFO.countries.map((country, index) => (
-        <span>
-          <span class="dashed-highlight">
-            {country}
+      <For each={TIMEZONE_INFO.countries}>
+        {(country, index) => (
+          <span>
+            <span class="dashed-highlight">
+              {country}
+            </span>
+            {index() < TIMEZONE_INFO.countries.length - 1 && (
+              index() === TIMEZONE_INFO.countries.length - 2 ? " and " : ", "
+            )}
           </span>
-          {index < TIMEZONE_INFO.countries.length - 1 && (
-            index === TIMEZONE_INFO.countries.length - 2 ? " and " : ", "
-          )}
-        </span>
-      ))}
+        )}
+      </For>
       . Right now it is{" "}
-      {props.showAnimation ? (
-        <ClientMotion
-          as="span"
-          class="dashed-highlight inline-block"
-          initial={{ opacity: 0 }}
-          animate={{ opacity: 1 }}
-          transition={{ duration: 0.6, delay: props.animationDelay || 0.8, easing: "ease-out" }}
-        >
-          <span class="font-mono leading-none">{currentTime()}</span>
-        </ClientMotion>
-      ) : (
-        <span class="dashed-highlight inline-block">
-          <span class="font-mono leading-none">{currentTime()}</span>
-        </span>
-      )}
+      <span class="dashed-highlight inline-block">
+        <span class="font-mono leading-none">{currentTime()}</span>
+      </span>
       .
-    </Motion.p>
+    </p>
   );
 }
